[feat] Add unit tests
