+++ hugegraph-loader/src/test/java/org/apache/hugegraph/loader/test/unit/TaskManagerTest.java
@@ -17,8 +17,11 @@
 
 package org.apache.hugegraph.loader.test.unit;
 
+import java.io.ByteArrayOutputStream;
+import java.io.PrintStream;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -48,7 +51,162 @@ import org.junit.Test;
 
 import org.apache.hugegraph.testutil.Assert;
 
-public class TaskManagerFailureTest {
+public class TaskManagerTest {
+
+    @Test
+    public void testBatchInsertFailureWithFallbackDisabled() throws Exception {
+        LoadOptions options = new LoadOptions();
+        options.batchFailureFallback = false;
+        Assert.assertFalse(options.batchFailureFallback);
+
+        LoadContext context = newTestContext(options);
+        TaskManager taskManager = new TaskManager(context);
+
+        EdgeMapping mapping = new EdgeMapping(Arrays.asList("s"), false,
+                                              Arrays.asList("t"), false);
+        mapping.label("knows");
+
+        InputStruct struct = new InputStruct(new ArrayList<>(),
+                                             new ArrayList<>());
+        struct.id("1");
+        struct.add(mapping);
+
+        LoadSummary summary = context.summary();
+        summary.inputMetricsMap()
+               .put(struct.id(), new LoadMetrics(struct));
+        LoadMetrics metrics = summary.metrics(struct);
+
+        setField(context.client(), "graph", newFailingBatchGraphManager());
+
+        List<Record> batch = new ArrayList<>();
+        batch.add(new Record("line1", new Edge("knows")));
+        batch.add(new Record("line2", new Edge("knows")));
+
+        ByteArrayOutputStream errOutput = new ByteArrayOutputStream();
+        PrintStream originalErr = System.err;
+        System.setErr(new PrintStream(errOutput, true,
+                                      StandardCharsets.UTF_8.name()));
+        try {
+            taskManager.submitBatch(struct, mapping, batch);
+            taskManager.waitFinished();
+
+            Assert.assertEquals(0L, flightingCount(metrics));
+            Assert.assertTrue(context.stopped());
+            Assert.assertFalse(context.noError());
+
+            String errText = errOutput.toString(StandardCharsets.UTF_8.name());
+            Assert.assertTrue(errText.contains(
+                    "Batch insert edges failed, stop loading."));
+
+            long before = flightingCount(metrics);
+            taskManager.submitBatch(struct, mapping, batch);
+            taskManager.waitFinished();
+            Assert.assertEquals(before, flightingCount(metrics));
+        } finally {
+            System.setErr(originalErr);
+            taskManager.shutdown();
+        }
+    }
+
+    @Test
+    public void testBatchInsertFailureWithFallbackEnabled() throws Exception {
+        LoadOptions options = new LoadOptions();
+        options.batchFailureFallback = true;
+
+        LoadContext context = newTestContext(options);
+        TaskManager taskManager = new TaskManager(context);
+
+        EdgeMapping mapping = new EdgeMapping(Arrays.asList("s"), false,
+                                              Arrays.asList("t"), false);
+        mapping.label("knows");
+
+        InputStruct struct = new InputStruct(new ArrayList<>(),
+                                             new ArrayList<>());
+        struct.id("1");
+        struct.add(mapping);
+
+        LoadSummary summary = context.summary();
+        summary.inputMetricsMap()
+               .put(struct.id(), new LoadMetrics(struct));
+        LoadMetrics metrics = summary.metrics(struct);
+
+        FailingBatchGraphManager.BATCH_CALLS.set(0);
+        FailingBatchGraphManager.SINGLE_CALLS.set(0);
+        setField(context.client(), "graph", newFailingBatchGraphManager());
+
+        List<Record> batch = new ArrayList<>();
+        batch.add(new Record("line1", new Edge("knows")));
+        batch.add(new Record("line2", new Edge("knows")));
+
+        try {
+            taskManager.submitBatch(struct, mapping, batch);
+            taskManager.waitFinished();
+
+            Assert.assertEquals(1, FailingBatchGraphManager.BATCH_CALLS.get());
+            Assert.assertEquals(2, FailingBatchGraphManager.SINGLE_CALLS.get());
+            Assert.assertEquals(0L, flightingCount(metrics));
+            Assert.assertFalse(context.stopped());
+            Assert.assertTrue(context.noError());
+        } finally {
+            taskManager.shutdown();
+        }
+    }
+
+    @Test
+    public void testMultipleBatchFailuresCounterConsistency() throws Exception {
+        LoadOptions options = new LoadOptions();
+        options.batchFailureFallback = true;
+
+        LoadContext context = newTestContext(options);
+        TaskManager taskManager = new TaskManager(context);
+
+        EdgeMapping mapping = new EdgeMapping(Arrays.asList("s"), false,
+                                              Arrays.asList("t"), false);
+        mapping.label("knows");
+
+        InputStruct struct = new InputStruct(new ArrayList<>(),
+                                             new ArrayList<>());
+        struct.id("1");
+        struct.add(mapping);
+
+        LoadSummary summary = context.summary();
+        summary.inputMetricsMap()
+               .put(struct.id(), new LoadMetrics(struct));
+        LoadMetrics metrics = summary.metrics(struct);
+
+        FailingBatchGraphManager.BATCH_CALLS.set(0);
+        FailingBatchGraphManager.SINGLE_CALLS.set(0);
+        setField(context.client(), "graph", newFailingBatchGraphManager());
+
+        List<Record> batch1 = new ArrayList<>();
+        batch1.add(new Record("line1", new Edge("knows")));
+        batch1.add(new Record("line2", new Edge("knows")));
+
+        List<Record> batch2 = new ArrayList<>();
+        batch2.add(new Record("line3", new Edge("knows")));
+        batch2.add(new Record("line4", new Edge("knows")));
+
+        try {
+            taskManager.submitBatch(struct, mapping, batch1);
+            taskManager.submitBatch(struct, mapping, batch2);
+            taskManager.waitFinished();
+
+            Assert.assertEquals(2, FailingBatchGraphManager.BATCH_CALLS.get());
+            Assert.assertEquals(4, FailingBatchGraphManager.SINGLE_CALLS.get());
+            Assert.assertEquals(0L, flightingCount(metrics));
+            Assert.assertFalse(context.stopped());
+            Assert.assertTrue(context.noError());
+
+            int expectedBatchPermits = 1 + options.batchInsertThreads;
+            int expectedSinglePermits = 2 * options.singleInsertThreads;
+            Assert.assertEquals(expectedBatchPermits,
+                                getSemaphorePermits(taskManager, "batchSemaphore"));
+            Assert.assertEquals(expectedSinglePermits,
+                                getSemaphorePermits(taskManager, "singleSemaphore"));
+        } finally {
+            taskManager.shutdown();
+        }
+    }
 
     @Test
     public void testConcurrentSubmitWhenStopping() throws Exception {
@@ -276,6 +434,10 @@ public class TaskManagerFailureTest {
         return (GraphManager) allocateInstance(FailingConcurrentGraphManager.class);
     }
 
+    private static GraphManager newFailingBatchGraphManager() throws Exception {
+        return (GraphManager) allocateInstance(FailingBatchGraphManager.class);
+    }
+
     private static GraphManager newSimpleGraphManager() throws Exception {
         return (GraphManager) allocateInstance(SimpleGraphManager.class);
     }
@@ -343,6 +505,31 @@ public class TaskManagerFailureTest {
             } catch (InterruptedException ignored) {
                 // Let the task finish on interruption.
             }
+        }
+    }
+
+    private static final class FailingBatchGraphManager extends GraphManager {
+
+        private static final AtomicInteger BATCH_CALLS = new AtomicInteger();
+        private static final AtomicInteger SINGLE_CALLS = new AtomicInteger();
+
+        private FailingBatchGraphManager() {
+            super(null, null, null);
+        }
+
+        @Override
+        public List<Edge> addEdges(List<Edge> edges, boolean checkVertex) {
+            return this.addEdges(edges);
+        }
+
+        @Override
+        public List<Edge> addEdges(List<Edge> edges) {
+            if (edges.size() > 1) {
+                BATCH_CALLS.incrementAndGet();
+                throw new RuntimeException("batch insert failure");
+            }
+            SINGLE_CALLS.addAndGet(edges.size());
+            return edges;
         }
     }
 }
